Clean deug info
