+++ domain/DeleteManifest.go
@@ -1,0 +1,28 @@
+/*
+Copyright 2021 Adobe. All rights reserved.
+This file is licensed to you under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License. You may obtain a copy
+of the License at http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software distributed under
+the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+OF ANY KIND, either express or implied. See the License for the specific language
+governing permissions and limitations under the License.
+*/
+
+package domain
+
+type Delete struct {
+	Name                 string `yaml:"name" json:"name"`
+	Type                 string `yaml:"type,omitempty" json:"type,omitempty"`
+	RefId                string `yaml:"refId,omitempty" json:"refId,omitempty"`
+	RequisiteStageRefIds []int  `yaml:"requisiteStageRefIds" json:"requisiteStageRefIds"`
+
+	App      string `yaml:"-" json:"app,omitempty"`
+	Location string `yaml:"-" json:"location,omitempty"`
+}
+
+func (del *Delete) DeleteManifest(p *Pipeline) {
+	del.App = p.Metadata.Application
+	del.Location = p.Namespace
+}
