+++ domain/pipeline.go
@@ -13,7 +13,6 @@ governing permissions and limitations under the License.
 package domain
 
 import (
-	"fmt"
 	"github.com/mitchellh/mapstructure"
 	log "github.com/sirupsen/logrus"
 )
@@ -22,7 +21,6 @@ type Pipeline struct {
 	PipelineManifest
 	PipelineSpec
 	Bake
-	Deploy
 }
 
 const (
@@ -34,17 +32,11 @@ const (
 
 func (p *Pipeline) ExpandSpec() {
 	for i := 0; i < len(p.Spec.Stages); i++ {
-		//stage := &p.Spec.Stages[i]
-		//stage.RefId = strconv.Itoa(i + 1)
-		//log.Debugf("Running stage: %v, RefId: %v", i, stage.RefId)
 		stage := new(Stage)
 		err := mapstructure.Decode(&p.Spec.Stages[i], stage)
 		if err != nil {
 			log.Fatalf("err: %v", err)
 		}
-		//fmt.Println(stage)
-		//fmt.Println(stage.Type)
-		err = mapstructure.Decode(&p.Spec.Stages[i], stage)
 		switch stage.Type {
 		case BakeManifest:
 			bake := new(Bake)
@@ -52,38 +44,26 @@ func (p *Pipeline) ExpandSpec() {
 			if err != nil {
 				log.Fatalf("err: %v", err)
 			}
-			fmt.Println("---------")
-			fmt.Println(bake.Type)
-			fmt.Println("---------")
-			bake.bakeManifest()
-			fmt.Println(bake.RequisiteStageRefIds)
-			fmt.Println("---------")
-			//stageList := make([]StageList, 0)
-			//stageList := append(bake)
+			bake.BakeManifest()
 		case deployManifest:
 			deploy := new(Deploy)
-			fmt.Println("deploy spec")
-			fmt.Println(&p.Spec.Stages[i])
-			fmt.Println("---------")
 			err = mapstructure.Decode(&p.Spec.Stages[i], deploy)
 			if err != nil {
 				log.Fatalf("err: %v", err)
 			}
-			fmt.Println("after decode")
-			fmt.Println(deploy.Type)
-			fmt.Println(deploy.RequisiteStageRefIds)
-			fmt.Println("---------")
 			deploy.DeployManifest(p)
 		case deleteManifest:
-			//stage.deleteManifest(p)
+			deleteStage := new(Delete)
+			err = mapstructure.Decode(&p.Spec.Stages[i], deleteStage)
+			if err != nil {
+				log.Fatalf("err: %v", err)
+			}
+			deleteStage.DeleteManifest(p)
 		default:
 		}
 	}
-	//os.Exit(0)
 }
 
-type StageList struct {
-}
 
 type Graph struct {
 	stages []*Stage
